refactor: rename InputCurrency to InputMoney (#59)
Purely mechanical rename aligning the money-amount input component with
the domain glossary: Money is the amount, currency stays as the unit
prop. Renames the component directory, input-group variant, test file,
and all consumer imports. No behavior, prop, or markup changes.
