--- src/lib/components/ui/input-currency/index.ts
@@ -1,3 +1,0 @@
-import Root from './input-currency.svelte';
-
-export { Root as InputCurrency };
