+++ src/lib/components/ui/input-group/index.ts
@@ -1,6 +1,6 @@
 import Addon from './input-group-addon.svelte';
 import Button from './input-group-button.svelte';
-import InputCurrency from './input-group-input-currency.svelte';
+import InputMoney from './input-group-input-money.svelte';
 import Input from './input-group-input.svelte';
 import Text from './input-group-text.svelte';
 import Textarea from './input-group-textarea.svelte';
@@ -10,15 +10,15 @@ export {
 	Addon,
 	Button,
 	Input,
-	InputCurrency,
 	//
 	Root as InputGroup,
 	Addon as InputGroupAddon,
 	Button as InputGroupButton,
 	Input as InputGroupInput,
-	InputCurrency as InputGroupInputCurrency,
+	InputMoney as InputGroupInputMoney,
 	Text as InputGroupText,
 	Textarea as InputGroupTextarea,
+	InputMoney,
 	Root,
 	Text,
 	Textarea
