+++ src/lib/components/ui/input-money/index.ts
@@ -1,0 +1,3 @@
+import Root from './input-money.svelte';
+
+export { Root as InputMoney };
